fix: Adds retry on conflict to STS updates.
When some changes occur to items that will force a STS update under the hood, there may become a version mismatch
between the k8 state and the state in our controller. Adding this retry logic will cause the STS to be refetched
prior to making the actual changes to STS that we want to.

fixes #41